requirements.txt
numpy
pandas
pytorch
matplotlib
